+++ packages/utility/distribution/src/Utility_DiscreteDistribution.hpp
@@ -102,7 +102,7 @@ public:
   IndepQuantity sampleAndRecordTrials( DistributionTraits::Counter& trials ) const override;
 
   //! Return a random sample and sampled index from the distribution
-  IndepQuantity sampleAndRecordBinIndex( unsigned& sampled_bin_index ) const override;
+  IndepQuantity sampleAndRecordBinIndex( size_t& sampled_bin_index ) const override;
 
   //! Return a random sample from the distribution at the given CDF value
   IndepQuantity sampleWithRandomNumber( const double random_number ) const override;
@@ -159,7 +159,7 @@ private:
 
   // Return a random sample using the random number and record the bin index
   IndepQuantity sampleImplementation( double random_number,
-				      unsigned& sampled_bin_index ) const;
+				      size_t& sampled_bin_index ) const;
 
   // Initialize the distribution
   void initializeDistribution(
